Refactor lading design
